drop table orders and timebased currency orders
